Make testbuild flexibler.

+++ oss-api/testbuild.sh
@@ -12,6 +12,14 @@ if [ "$1" ]; then
 	scp -P $PORT de.openschoolserver.api-1.0-SNAPSHOT/lib/de.openschoolserver.* root@$1:/opt/oss-java/lib/
 	ssh -p $PORT root@$1 systemctl restart oss-api
 fi
+echo  -n "Do you want to check in (y/n)?"
+read Y
+if [ "$Y" != "y" ]; then
+	exit
+fi
+cd /data1/OSC/home:varkoly:OSS-4-0/oss-java/
+osc up
+cd $HERE
 if [ -e oss-java ]; then
     rm -r oss-java
 fi
@@ -24,4 +32,9 @@ cp ../oss-dao/data/business-INSERT.sql /data1/OSC/home:varkoly:OSS-4-0/oss-java/
 CLASSPATH=$( grep "^CLASSPATH=" oss-java/bin/de.openschoolserver.api )
 sed "s#@CLASSPATH@#$CLASSPATH#" start-oss-api > /data1/OSC/home:varkoly:OSS-4-0/oss-java/start-oss-api
 cp start-oss-squid-sso /data1/OSC/home:varkoly:OSS-4-0/oss-java/start-oss-squid-sso
-#rm -r oss-java
+rm -r oss-java
+xterm -e git log --raw  &
+cd /data1/OSC/home:varkoly:OSS-4-0/oss-java/
+osc vc
+osc ci
+cd $HERE
